Creates component for breakout invitation
