+++ bigbluebutton-html5/imports/ui/components/app/component.jsx
@@ -9,6 +9,7 @@ import PollingContainer from '/imports/ui/components/polling/container';
 import logger from '/imports/startup/client/logger';
 import ActivityCheckContainer from '/imports/ui/components/activity-check/container';
 import UserInfoContainer from '/imports/ui/components/user-info/container';
+import BreakoutRoomInvitation from '/imports/ui/components/breakout-room/invitation/container';
 import ToastContainer from '../toast/container';
 import ModalContainer from '../modal/container';
 import NotificationsBarContainer from '../notifications-bar/container';
@@ -289,6 +290,7 @@ class App extends Component {
           {this.renderPanel()}
           {this.renderSidebar()}
         </section>
+        <BreakoutRoomInvitation />
         <PollingContainer />
         <ModalContainer />
         <AudioContainer />
