+++ bigbluebutton-html5/imports/ui/components/breakout-join-confirmation/container.jsx
@@ -3,7 +3,7 @@ import { withTracker } from 'meteor/react-meteor-data';
 import Breakouts from '/imports/api/breakouts';
 import Auth from '/imports/ui/services/auth';
 import { makeCall } from '/imports/ui/services/api';
-import navBarService from '/imports/ui/components/nav-bar/service';
+import breakoutService from '/imports/ui/components/breakout-room/service';
 import BreakoutJoinConfirmationComponent from './component';
 
 const BreakoutJoinConfirmationContrainer = props =>
@@ -36,7 +36,7 @@ export default withTracker(({ breakout, mountModal, breakoutName }) => {
     mountModal,
     breakoutName,
     breakoutURL: url,
-    breakouts: navBarService.getBreakouts(),
+    breakouts: breakoutService.getBreakouts(),
     requestJoinURL,
     getURL,
   };
