+++ bigbluebutton-html5/imports/ui/components/nav-bar/container.jsx
@@ -56,7 +56,6 @@ export default withTracker(() => {
     },
   });
 
-  const breakouts = Service.getBreakouts();
   const currentUserId = Auth.userID;
   const { connectRecordingObserver, processOutsideToggleRecording } = Service;
   const currentUser = Users.findOne({ userId: Auth.userID });
@@ -67,7 +66,6 @@ export default withTracker(() => {
   return {
     amIModerator,
     isExpanded,
-    breakouts,
     currentUserId,
     processOutsideToggleRecording,
     connectRecordingObserver,
@@ -75,8 +73,6 @@ export default withTracker(() => {
     presentationTitle: meetingTitle,
     hasUnreadMessages: checkUnreadMessages(),
     isBreakoutRoom: meetingIsBreakout(),
-    getBreakoutByUser: Service.getBreakoutByUser,
-    currentBreakoutUser: Service.getBreakoutUserByUserId(Auth.userID),
     recordProps: meetingRecorded,
     toggleUserList: () => {
       Session.set('isUserListOpen', !isExpanded);
