+++ bigbluebutton-html5/imports/ui/components/nav-bar/service.js
@@ -1,31 +1,6 @@
 import Auth from '/imports/ui/services/auth';
-import Breakouts from '/imports/api/breakouts';
 import { makeCall } from '/imports/ui/services/api';
 import Meetings from '/imports/api/meetings';
-import fp from 'lodash/fp';
-
-const getBreakoutByUserId = userId => Breakouts.find({ 'users.userId': userId }).fetch();
-
-const getBreakoutByUser = user => Breakouts.findOne({ users: user });
-
-const getUsersFromBreakouts = breakoutsArray => breakoutsArray
-  .map(breakout => breakout.users)
-  .reduce((acc, usersArray) => [...acc, usersArray], []);
-
-const filterUserURLs = userId => breakoutUsersArray => breakoutUsersArray
-  .filter(user => user.userId === userId);
-
-const getLastURLInserted = breakoutURLArray => breakoutURLArray
-  .sort((a, b) => a.insertedTime - b.insertedTime).pop();
-
-const getBreakoutUserByUserId = userId => fp.pipe(
-  getBreakoutByUserId,
-  getUsersFromBreakouts,
-  filterUserURLs(userId),
-  getLastURLInserted,
-)(userId);
-
-const getBreakouts = () => Breakouts.find({}, { sort: { sequence: 1 } }).fetch();
 
 const processOutsideToggleRecording = (e) => {
   switch (e.data) {
@@ -45,7 +20,6 @@ const processOutsideToggleRecording = (e) => {
   }
 };
 
-
 const connectRecordingObserver = () => {
   // notify on load complete
   this.window.parent.postMessage({ response: 'readyToConnect' }, '*');
@@ -54,7 +28,4 @@ const connectRecordingObserver = () => {
 export default {
   connectRecordingObserver: () => connectRecordingObserver(),
   processOutsideToggleRecording: arg => processOutsideToggleRecording(arg),
-  getBreakoutUserByUserId,
-  getBreakoutByUser,
-  getBreakouts,
 };
