+++ bigbluebutton-html5/imports/ui/components/nav-bar/styles.scss
@@ -82,7 +82,3 @@
 
 .btnSettings {
 }
-
-.dropdownBreakout {
-  cursor: pointer;
-}
